removed duplicate creation test logic
