+++ mediator/src/routes/encounter.ts
@@ -1,8 +1,7 @@
 import { Router } from 'express';
 import { validateBodyAgainst } from '../middlewares';
 import { EncounterSchema } from '../middlewares/schemas/encounter';
-import { createEncounter } from '../controllers/encounter';
-import { validateFhirResource } from '../utils/fhir';
+import { createFhirResource, validateFhirResource } from '../utils/fhir';
 import { requestHandler } from '../utils/request';
 
 const router = Router();
@@ -12,7 +11,7 @@ const resourceType = 'Encounter';
 router.post(
   '/',
   validateBodyAgainst(validateFhirResource(resourceType), EncounterSchema),
-  requestHandler((req) => createEncounter({ ...req.body, resourceType }))
+  requestHandler((req) => createFhirResource({ ...req.body, resourceType }))
 );
 
 export default router;
