+++ mediator/src/routes/endpoint.ts
@@ -1,8 +1,7 @@
 import { Router } from 'express';
 import { validateBodyAgainst } from '../middlewares';
-import { validateFhirResource } from '../utils/fhir';
+import { createFhirResource, validateFhirResource } from '../utils/fhir';
 import { requestHandler } from '../utils/request';
-import { createEndpoint } from '../controllers/endpoint';
 
 const router = Router();
 
@@ -11,7 +10,7 @@ const resourceType = 'Endpoint';
 router.post(
   '/',
   validateBodyAgainst(validateFhirResource(resourceType)),
-  requestHandler((req) => createEndpoint({ ...req.body, resourceType }))
+  requestHandler((req) => createFhirResource({ ...req.body, resourceType }))
 );
 
 export default router;
