+++ mediator/src/routes/organization.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express';
-import { createOrganization } from '../controllers/organization';
 import { validateBodyAgainst } from '../middlewares';
-import { validateFhirResource } from '../utils/fhir';
+import { createFhirResource, validateFhirResource } from '../utils/fhir';
 import { OrganizationSchema } from '../middlewares/schemas/organization';
 import { requestHandler } from '../utils/request';
 
@@ -12,7 +11,7 @@ const resourceType = 'Organization';
 router.post(
   '/',
   validateBodyAgainst(validateFhirResource(resourceType), OrganizationSchema),
-  requestHandler((req) => createOrganization({ ...req.body, resourceType }))
+  requestHandler((req) => createFhirResource({ ...req.body, resourceType }))
 );
 
 export default router;
